Memory: Further leaks fixed.
All EN leaks save physics *phy removed.
Looks like two in epoxy still.
